almost_crud_2_no_transactions

+++ lib/Freee/Helpers/PgSettings.pm
@@ -229,31 +229,32 @@ sub register {
     ###################################################################
 
 
-    # для создания возможностей пользователя
+    # для создания возможностей групп пользователей
     # my $id = $self->insert_group({
-    #     "folder"      => 0,           - это возможности пользователя
-    #     "lib_id"      => 0,           - обязательно (должно быть натуральным числом)
+    #       "folder"      => 0,           - это возможности пользователя
+    #     "lib_id"      => 5,           - обязательно id родителя (должно быть натуральным числом)
     #     "label"       => 'название',  - обязательно (название для отображения)
     #     "name",       => 'name'       - обязательно (системное название, латиница)
-    #     "editable"    => 0,           - не обязательно, по умолчанию 1
+    #     "editable"    => 0,           - не обязательно, по умолчанию 0
     #     "readOnly"    => 0,           - не обязательно, по умолчанию 0
-    #     "removable"   => 0,           - не обязательно, по умолчанию 1
+    #     "removable"   => 0,           - не обязательно, по умолчанию 0
     #     "value"       => "",            - строка или json
-    #     "required"    => 0              - обязательное поле
+    #     "required"    => 0            - не обязательно, по умолчанию 0
     # });
     # для создания группы пользователей
     # my $id = $self->insert_group({
-    #     "folder"      => 1,           - это пользователь
-    #     "lib_id"      => 0,           - обязательно (должно быть натуральным числом)
+    #     "folder"      => 1,           - это группа пользователей
+    #     "lib_id"      => 0,           - обязательно 0 (должно быть натуральным числом) 
     #     "label"       => 'название',  - обязательно (название для отображения)
     #     "name",       => 'name'       - обязательно (системное название, латиница)
-    #     "editable"    => 0,           - не обязательно, по умолчанию 1
+    #     "editable"    => 0,           - не обязательно, по умолчанию 0
     #     "readOnly"    => 0,           - не обязательно, по умолчанию 0
-    #     "removable"   => 0,           - не обязательно, по умолчанию 1
-    # });    
+    #     "removable"   => 0,           - не обязательно, по умолчанию 0
+    # });
+    # возвращается id записи    
     $app->helper( 'insert_group' => sub {
         my ($self, $data) = @_;
-        #return unless $data;
+        return unless $data;
 
         $self->pg_dbh->do('INSERT INTO "public"."groups" ('.join( ',', map { "\"$_\""} keys %$data ).') VALUES ('.join( ',', map { $self->pg_dbh->quote( $$data{$_} ) } keys %$data ).') RETURNING "id"');
         my $id = $self->pg_dbh->last_insert_id(undef, 'public', 'groups', undef, { sequence => 'groups_id_seq' });
@@ -262,17 +263,37 @@ sub register {
     });
 
 
-    # для редактирования группы
-    # возвращается id записи
+    # для создания возможностей групп пользователей
+    # my $id = $self->insert_group({
+    #     "folder"      => 0,           - это возможности пользователя
+    #     "lib_id"      => 5,           - обязательно id родителя (должно быть натуральным числом)
+    #     "label"       => 'название',  - обязательно (название для отображения)
+    #     "name",       => 'name'       - обязательно (системное название, латиница)
+    #     "editable"    => 0,           - не обязательно, по умолчанию 0
+    #     "readOnly"    => 0,           - не обязательно, по умолчанию 0
+    #     "removable"   => 0,           - не обязательно, по умолчанию 0
+    #     "value"       => "",            - строка или json
+    #     "required"    => 0            - не обязательно, по умолчанию 0
+    # });
+    # для создания группы пользователей
+    # my $id = $self->insert_group({
+    #     "folder"      => 1,           - это группа пользователей
+    #     "lib_id"      => 0,           - обязательно 0 (должно быть натуральным числом) 
+    #     "label"       => 'название',  - обязательно (название для отображения)
+    #     "name",       => 'name'       - обязательно (системное название, латиница)
+    #     "editable"    => 0,           - не обязательно, по умолчанию 0
+    #     "readOnly"    => 0,           - не обязательно, по умолчанию 0
+    #     "removable"   => 0,           - не обязательно, по умолчанию 0
+    # });
+    # возвращается true/false
     $app->helper( 'update_group' => sub {
         my ($self, $data) = @_;
-        #return unless $data;
+        return unless $data;
 
         my $db_result = $self->pg_dbh->do('UPDATE "public"."groups" SET '.join( ', ', map { "\"$_\"=".$self->pg_dbh->quote( $$data{$_} ) } keys %$data )." WHERE \"id\"=".$self->pg_dbh->quote( $$data{id} )." RETURNING \"id\"") if $$data{id};
 
         return $db_result;
     });
-
 
 
     # для удаления группы пользователей
@@ -280,28 +301,109 @@ sub register {
     # возвращается true/false
     $app->helper( 'delete_group' => sub {
         my ($self, $id) = @_;
-        #return unless $id;
+        return unless $id;
 
+        # $dbh->begin_work();
+
+        # unless ( $self->pg_dbh->do('DELETE FROM "public"."groups" WHERE "lib_id"='.$id) ) {
+        #     $dbh->rollback();
+        # } 
+        # else {
+        #     unless ( $self->pg_dbh->do('DELETE FROM "public"."groups" WHERE "id"='.$id ) ) {
+        #         $dbh->rollback();
+        #     }
+        #     else {
+        #         $db_result = 1;
+        #     }
+        # }
+
+        # $dbh->commit();
+
+        # return $db_result;
+
+
+
+
+        # my $pg_dbh;
+        # $self->$pg_dbh->begin_work();
+        # my $db_result = $self->pg_dbh->do('DELETE FROM "public"."groups" WHERE "lib_id"='.$id);
+        # print Dumper($db_result);
+        # $db_result = 0;
+        # unless ( $db_result ) {
+        #     $self->$pg_dbh->rollback();
+        #     return 0;
+        # }
+        # $db_result = $self->pg_dbh->do('DELETE FROM "public"."groups" WHERE "id"='.$id);
+        # print Dumper($db_result);
+        # unless ( $db_result ) {
+        #     $self->$pg_dbh->rollback();
+        #     return 0;
+        # }
+        # $self->$pg_dbh->commit();
+        # return $db_result;
+        
+
+
+
+        
         my $db_result = $self->pg_dbh->do('DELETE FROM "public"."groups" WHERE "lib_id"='.$id);
         $db_result = $self->pg_dbh->do('DELETE FROM "public"."groups" WHERE "id"='.$id);
 
         return $db_result;
+
     });
+
 
     # для изменения параметра status
     # возвращается true/false
     $app->helper( 'status_group' => sub {
         my ($self, $data) = @_;
-        #return unless $data;
-
-        my $id = $self->param('id');
-        my $status = $self->param('status');
+        return unless $data;
 
         my $db_result = $self->pg_dbh->do('UPDATE "public"."groups" SET '.join( ', ', map { "\"$_\"=".$self->pg_dbh->quote( $$data{$_} ) } keys %$data )." WHERE \"id\"=".$self->pg_dbh->quote( $$data{id} )." RETURNING \"id\"") if $$data{id};
 
         return $db_result;
     });
 
+
+    # для проверки корректности наследования
+    # my $true = $self->lib_id_check( 99 );
+    # возвращается true/false
+    $app->helper( 'lib_id_check' => sub {
+        my ($self, $lib_id) = @_;
+
+        # это не фолдер?
+        if ( $lib_id ) {
+
+            # есть родитель?
+            if ( $self->id_check( $lib_id ) ) {
+
+                # родитель фолдер?
+                if ( $self->pg_dbh->selectrow_array( 'SELECT lib_id FROM "public"."groups" WHERE "id"='.$lib_id ) ) {
+                    return 0;
+                }
+            } 
+
+            else {
+                return 0;
+            }
+        }
+
+        return 1;
+    });
+
+
+    # для проверки существования строки с данным id
+    # my $true = $self->id_check( 99 );
+    # возвращается true/false
+    $app->helper( 'id_check' => sub {
+        my ($self, $id) = @_;
+        return unless $id;
+
+        my $db_result = $self->pg_dbh->selectrow_hashref('SELECT * FROM "public"."groups" WHERE "id"='.$id);
+
+        return $db_result;
+    });
 }
 
 1;
